project/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-JWT-Extended
Flask-Mail
Flask-SocketIO
Flask-Admin
Flask-Cors
psycopg2-binary
Pillow
pytesseract
opencv-python
reportlab
spacy
sentence-transformers
python-dotenv
requests
pandas
numpy
gunicorn
openai
